Add files via upload
Created index3,php file. Is a copy of index.php file, but now we can upload images to generate puzzles with images
In index3.php page, the user can choose letters  or images as puzzle inputs.
If images inputs, the user must upload images as the size of the puzzle just as in the letters version.
Since our puzzle cells are square, I wrote a function which will prompt the user to crop the image during upload. The image  are also resized.
Upon submission,  if it is an image puzzle, we  create a unique folder inside uploads folder and save the uploaded images to this unique folder. We follow the same code format, for the image puzzle we redirect to wordokuPuzzle3.php

Here we are not dealing with words. But images. I created wordoku3.php file in which I modify the wordoku class to suite my images puzzle. Most especially, no need to call $this->convertToLetters(); since we are not dealing with letters.
With the indexes in the puzzle and solution arrays, I was able to display the images instead of letters in each cell as necessary.